Added player, drew grid, added player movement, refactored code
Created a Player class with information related to the player character, such as position, health, and score. Displayed a square grid on the window that resizes automatically to fit the shortest dimension (either width or height). Incorporated player movement inside the grid, and made it so the player can't move outside of the grid boundaries. Refactored the code to split Main.py into a class that contained various functions rather than being just one large function.